make

+++ hexa.c
@@ -6,7 +6,7 @@
 /*   By: mfrasson <mfrasson@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/21 20:55:25 by mfrasson          #+#    #+#             */
-/*   Updated: 2021/03/21 22:23:49 by mfrasson         ###   ########.fr       */
+/*   Updated: 2021/03/21 22:24:47 by mfrasson         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ void		print_flag_p(va_list args, t_flags *flag, t_counter *count)
 	str = ft_itoa_hex(count, va_arg(args, unsigned long int));
 	if (!(*str == '0' && flag->precision == 0))
     {
-        if (flag->precision < ft_strlen(str))
+        if (flag->precision < (int)ft_strlen(str))
             flag->precision = ft_strlen(str);
         if (flag->width < flag->precision + 2)
             flag->width = flag->precision + 2;
